+++ src/theme/alert.ts
@@ -29,28 +29,29 @@ export default {
   },
   variants: {
     color: {
-      'air-primary': { root: 'air-primary' },
-      'air-primary-success': { root: 'air-primary-success' },
-      'air-primary-alert': { root: 'air-primary-alert' },
-      'air-primary-copilot': { root: 'air-primary-copilot' },
-      'air-secondary': { root: 'air-secondary' },
-      'air-secondary-alert': { root: 'air-secondary-alert' },
-      'air-secondary-accent': { root: 'air-secondary-accent' },
-      'air-secondary-accent-1': { root: 'air-secondary-accent-1' },
-      'air-secondary-accent-2': { root: 'air-secondary-accent-2' },
-      'air-secondary-no-accent': { root: 'air-secondary-no-accent' },
-      'air-tertiary': { root: 'air-tertiary' },
-      'air-tertiary-accent': { root: 'air-tertiary-accent' },
-      'air-tertiary-no-accent': { root: 'air-tertiary-no-accent' },
+      'air-primary': { root: 'style-filled' },
+      'air-primary-success': { root: 'style-filled-success' },
+      'air-primary-alert': { root: 'style-filled-alert' },
+      'air-primary-copilot': { root: 'style-filled-copilot' },
+      'air-primary-warning': { root: 'style-filled-warning' },
+      'air-secondary': { root: 'style-tinted' },
+      'air-secondary-alert': { root: 'style-tinted-alert' },
+      'air-secondary-accent': { root: 'style-outline' },
+      'air-secondary-accent-1': { root: 'style-outline-accent-1' },
+      'air-secondary-accent-2': { root: 'style-outline-accent-2' },
+      'air-secondary-no-accent': { root: 'style-outline-no-accent' },
+      'air-tertiary': { root: 'style-plain' },
+      'air-tertiary-accent': { root: 'style-plain-accent' },
+      'air-tertiary-no-accent': { root: 'style-plain-no-accent' },
       // @deprecate ////
-      'default': { root: 'old-style-default' },
-      'danger': { root: 'old-style-danger' },
-      'success': { root: 'old-style-success' },
-      'warning': { root: 'old-style-warning' },
-      'primary': { root: 'old-style-primary' },
-      'secondary': { root: 'old-style-secondary' },
-      'collab': { root: 'old-style-collab' },
-      'ai': { root: 'old-style-ai' }
+      'default': { root: 'style-old-default' },
+      'danger': { root: 'style-old-danger' },
+      'success': { root: 'style-old-success' },
+      'warning': { root: 'style-old-warning' },
+      'primary': { root: 'style-old-primary' },
+      'secondary': { root: 'style-old-secondary' },
+      'collab': { root: 'style-old-collab' },
+      'ai': { root: 'style-old-ai' }
     },
     size: {
       sm: {
@@ -80,9 +81,51 @@ export default {
       true: {
         description: 'mt-1'
       }
+    },
+    inverted: {
+      true: '',
+      false: ''
     }
   },
-  compoundVariants: [],
+  compoundVariants: [
+    // region color ////
+    {
+      inverted: true,
+      color: 'air-primary',
+      class: {
+        root: 'style-filled-inverted'
+      }
+    },
+    {
+      inverted: true,
+      color: 'air-primary-success',
+      class: {
+        root: 'style-filled-success-inverted'
+      }
+    },
+    {
+      inverted: true,
+      color: 'air-primary-alert',
+      class: {
+        root: 'style-filled-alert-inverted'
+      }
+    },
+    {
+      inverted: true,
+      color: 'air-primary-copilot',
+      class: {
+        root: 'style-filled-copilot-inverted'
+      }
+    },
+    {
+      inverted: true,
+      color: 'air-primary-warning',
+      class: {
+        root: 'style-filled-warning-inverted'
+      }
+    }
+    // endregion ////
+  ],
   defaultVariants: {
     color: 'air-secondary-no-accent',
     size: 'md'